# swagger.yaml
definitions:
    Product:
        description: Product defines the structure for an API product
        properties:
            description:
                type: string
                x-go-name: Description
            id:
                format: int64
                type: integer
                x-go-name: ID
            name:
                type: string
                x-go-name: Name
            price:
                format: float
                type: number
                x-go-name: Price
            sku:
                type: string
                x-go-name: SKU
        type: object
        x-go-package: _/home/sahas/documents/projects/go-micro/data
paths:
    /products:
        get:
            description: Return a list of products from the database
            operationId: listProducts
            responses:
                "200":
                    $ref: '#/responses/productsResponse'
            tags:
                - products
        post:
            description: Create a new Product
            operationId: addProducts
            responses:
                "200":
                    $ref: '#/responses/productResponse'
                "422":
                    $ref: '#/responses/errorValidation'
                "501":
                    $ref: '#/responses/errorResponse'
            tags:
                - products
    /products/{id}:
        delete:
            description: Update a products details
            operationId: deleteProduct
            responses:
                "201":
                    $ref: '#/responses/noContentResponse'
                "404":
                    $ref: '#/responses/errorResponse'
                "501":
                    $ref: '#/responses/errorResponse'
            tags:
                - products
        get:
            description: Return a list of products from the database
            operationId: listSingle
            responses:
                "200":
                    $ref: '#/responses/productResponse'
                "404":
                    $ref: '#/responses/errorResponse'
            tags:
                - products
swagger: "2.0"